AST-702: Move CustomerHeaderStamp to Component directory
